RetroPlayer: Separate GUI bridge code
And add some much-needed documentation.
