+++ xbmc/cores/RetroPlayer/RetroPlayer.cpp
@@ -25,8 +25,8 @@
 #include "RetroPlayerVideo.h"
 #include "addons/AddonManager.h"
 #include "cores/DataCacheCore.h"
+#include "cores/RetroPlayer/guibridge/GUIGameRenderManager.h"
 #include "cores/RetroPlayer/process/RPProcessInfo.h"
-#include "cores/RetroPlayer/rendering/GUIGameRenderManager.h"
 #include "cores/RetroPlayer/rendering/RPRenderManager.h"
 #include "dialogs/GUIDialogYesNo.h"
 #include "filesystem/File.h"
