+++ xbmc/cores/RetroPlayer/guibridge/GUIGameSettings.cpp
