+++ xbmc/cores/RetroPlayer/guibridge/GUIGameVideoHandle.cpp
