+++ xbmc/cores/RetroPlayer/guibridge/GUIGameVideoHandle.h
