+++ xbmc/cores/RetroPlayer/guibridge/GUIRenderHandle.cpp
