+++ xbmc/cores/RetroPlayer/guibridge/GUIRenderHandle.h
