+++ xbmc/cores/RetroPlayer/guibridge/GUIRenderTargetFactory.cpp
