+++ xbmc/cores/RetroPlayer/guibridge/GUIRenderTargetFactory.h
