+++ xbmc/cores/RetroPlayer/guibridge/IRenderCallback.h
