+++ xbmc/cores/RetroPlayer/guiwindows/GameWindowFullScreen.cpp
@@ -20,8 +20,8 @@
 
 #include "GameWindowFullScreen.h"
 #include "GameWindowFullScreenText.h"
-#include "cores/RetroPlayer/rendering/GUIGameRenderManager.h"
-#include "cores/RetroPlayer/rendering/GUIRenderHandle.h"
+#include "cores/RetroPlayer/guibridge/GUIGameRenderManager.h"
+#include "cores/RetroPlayer/guibridge/GUIRenderHandle.h"
 #include "guilib/GraphicContext.h" //! @todo Remove me
 #include "guilib/GUIDialog.h"
 #include "guilib/GUIControl.h"
