+++ xbmc/cores/RetroPlayer/rendering/CMakeLists.txt
@@ -1,10 +1,4 @@
-set(SOURCES GUIGameRenderManager.cpp
-            GUIGameSettings.cpp
-            GUIGameVideoHandle.cpp
-            GUIRenderHandle.cpp
-            GUIRenderTarget.cpp
-            GUIRenderTargetFactory.cpp
-            RenderContext.cpp
+set(SOURCES RenderContext.cpp
             RenderGeometry.cpp
             RenderSettings.cpp
             RenderTranslator.cpp
@@ -12,15 +6,7 @@ set(SOURCES GUIGameRenderManager.cpp
             RPRenderManager.cpp
 )
 
-set(HEADERS GUIGameRenderManager.h
-            GUIGameSettings.h
-            GUIGameVideoHandle.h
-            GUIRenderHandle.h
-            GUIRenderTarget.h
-            GUIRenderTargetFactory.h
-            IGUIRenderSettings.h
-            IRenderCallback.h
-            IRenderManager.h
+set(HEADERS IRenderManager.h
             RenderContext.h
             RenderGeometry.h
             RenderSettings.h
