+++ xbmc/games/dialogs/osd/DialogGameVideoFilter.cpp
@@ -19,7 +19,7 @@
  */
 
 #include "DialogGameVideoFilter.h"
-#include "cores/RetroPlayer/rendering/GUIGameVideoHandle.h"
+#include "cores/RetroPlayer/guibridge/GUIGameVideoHandle.h"
 #include "guilib/LocalizeStrings.h"
 #include "guilib/WindowIDs.h"
 #include "settings/GameSettings.h"
